Extract songs fetch-lyrics core
